Split checkpoint sync helpers
