+++ .pm/tasks/task_f81d3e661b7048d8b2fe6987a544a368.execution.md
@@ -212,3 +212,12 @@ Example:
 - Expected Result: PR preflight accepts local role review evidence and opens a normal PR CI/watch path.
 - Actual Result: PR created: https://github.com/eng-cc/oasis7/pull/409. Preflight reported branch ahead of base by 2, behind by 0, no branch sync suggested, and recommended targeted required validation command.
 - Blocker / Next Action: Watch required checks, mergeability, comments, and review threads; after merge, run CI package and deploy validation.
+
+## 2026-06-11 23:12:00 CST / tpm
+- 完成内容: Fixed local required-gate blockers found after PR creation.
+- 遗留事项: GitHub checks need to rerun on the pushed fix; full local required claim had one unrelated viewer runtime_live timing failure before targeted rerun passed.
+- Action: Split checkpoint helper methods out of oversized `crates/oasis7_node/src/replication.rs` into `replication_checkpoint.rs` and `replication_sampling.rs`; split checkpoint driver tests from oversized `execution_bridge/tests/driver.rs` into `driver_checkpoint.rs`. This resolved the required `check-rust-file-size.sh` failure. Cleaned stale package artifacts for `oasis7_node`/`oasis7` after a stale test build reported missing checkpoint API that current `cargo check` did not reproduce.
+- Validation Command: `cargo fmt --check`; `git diff --check`; `./scripts/check-rust-file-size.sh`; `env -u RUSTC_WRAPPER cargo check -p oasis7_node -p oasis7 --bin oasis7_chain_runtime`; `env -u RUSTC_WRAPPER cargo test -p oasis7_node high_checkpoint -- --nocapture`; `env -u RUSTC_WRAPPER cargo test -p oasis7 --bin oasis7_chain_runtime execution_bridge -- --nocapture`; `cargo clean -p oasis7_node -p oasis7 && env -u RUSTC_WRAPPER cargo test -p oasis7 chain_linked_gameplay_action_submits_to_chain_and_applies_on_committed_sync --features test_tier_required -- --nocapture`.
+- Expected Result: Required file-size gate passes; split modules preserve checkpoint behavior; stale artifact issue is cleared; viewer runtime_live failure is not deterministic.
+- Actual Result: File-size gate passed with oversized code files=0 and test files=0. `cargo check` passed with only pre-existing dead-code warnings. `high_checkpoint` passed 7 tests. `oasis7_chain_runtime execution_bridge` passed 47 tests including the moved `driver_checkpoint` tests. The targeted viewer runtime_live rerun passed after package clean.
+- Blocker / Next Action: Commit/push split fix to PR #409 and watch GitHub checks.
