+++ crates/oasis7_node/src/replication.rs
@@ -1,13 +1,11 @@
 use crate::{
-    NodeConsensusAction, NodeError, NodeExecutionCheckpointBlob, NodeExecutionCheckpointBlobRef,
-    NodeExecutionCheckpointBundle, NodeExecutionCheckpointDescriptor, PosConsensusStatus,
-    PosDecision,
+    NodeConsensusAction, NodeError, NodeExecutionCheckpointBundle,
+    NodeExecutionCheckpointDescriptor, PosConsensusStatus, PosDecision,
 };
 use ed25519_dalek::{Signature, Signer, SigningKey, Verifier, VerifyingKey};
 use oasis7_distfs::{
     apply_replication_record, blake3_hex, build_replication_record_with_epoch, BlobStore as _,
     FileReplicationRecord, FileStore as _, LocalCasStore, SingleWriterReplicationGuard,
-    StorageChallengeProbeConfig, StorageChallengeProbeReport,
 };
 use oasis7_proto::world_error::WorldError;
 use serde::{Deserialize, Serialize};
@@ -25,6 +23,10 @@ pub(crate) const REPLICATION_FETCH_COMMIT_PROTOCOL: &str =
     "/aw/node/replication/fetch-commit/1.0.0";
 pub(crate) const REPLICATION_FETCH_BLOB_PROTOCOL: &str = "/aw/node/replication/fetch-blob/1.0.0";
 mod commit_retention;
+#[path = "replication_checkpoint.rs"]
+mod replication_checkpoint;
+#[path = "replication_sampling.rs"]
+mod replication_sampling;
 #[path = "replication_support.rs"]
 mod support;
 
@@ -1049,182 +1051,6 @@ impl ReplicationRuntime {
             request.requester_signature_hex = Some(sign_fetch_blob_request(&request, signer)?);
         }
         Ok(request)
-    }
-
-    pub(crate) fn store_execution_checkpoint_bundle(
-        &self,
-        bundle: &NodeExecutionCheckpointBundle,
-    ) -> Result<NodeExecutionCheckpointDescriptor, NodeError> {
-        let manifest_ref = blake3_hex(bundle.manifest_json.as_slice());
-        self.store_blob_by_hash(manifest_ref.as_str(), bundle.manifest_json.as_slice())?;
-        let mut blob_refs = Vec::with_capacity(bundle.blobs.len());
-        for blob in &bundle.blobs {
-            let actual = blake3_hex(blob.bytes.as_slice());
-            if actual != blob.content_hash {
-                return Err(NodeError::Replication {
-                    reason: format!(
-                        "execution checkpoint blob hash mismatch expected={} actual={}",
-                        blob.content_hash, actual
-                    ),
-                });
-            }
-            self.store_blob_by_hash(blob.content_hash.as_str(), blob.bytes.as_slice())?;
-            blob_refs.push(NodeExecutionCheckpointBlobRef {
-                content_hash: blob.content_hash.clone(),
-                size_bytes: blob.bytes.len() as u64,
-            });
-        }
-        let descriptor = NodeExecutionCheckpointDescriptor {
-            height: bundle.height,
-            execution_block_hash: bundle.execution_block_hash.clone(),
-            execution_state_root: bundle.execution_state_root.clone(),
-            manifest_ref,
-            manifest_size_bytes: bundle.manifest_json.len() as u64,
-            blobs: blob_refs,
-        };
-        self.pin_execution_checkpoint_descriptor(&descriptor)?;
-        Ok(descriptor)
-    }
-
-    pub(crate) fn load_execution_checkpoint_bundle(
-        &self,
-        descriptor: &NodeExecutionCheckpointDescriptor,
-    ) -> Result<Option<NodeExecutionCheckpointBundle>, NodeError> {
-        let Some(manifest_json) = self.load_blob_by_hash(descriptor.manifest_ref.as_str())? else {
-            return Ok(None);
-        };
-        if manifest_json.len() as u64 != descriptor.manifest_size_bytes {
-            return Err(NodeError::Replication {
-                reason: format!(
-                    "execution checkpoint manifest size mismatch expected={} actual={}",
-                    descriptor.manifest_size_bytes,
-                    manifest_json.len()
-                ),
-            });
-        }
-        let mut blobs = Vec::with_capacity(descriptor.blobs.len());
-        for blob_ref in &descriptor.blobs {
-            let Some(bytes) = self.load_blob_by_hash(blob_ref.content_hash.as_str())? else {
-                return Ok(None);
-            };
-            if bytes.len() as u64 != blob_ref.size_bytes {
-                return Err(NodeError::Replication {
-                    reason: format!(
-                        "execution checkpoint blob size mismatch hash={} expected={} actual={}",
-                        blob_ref.content_hash,
-                        blob_ref.size_bytes,
-                        bytes.len()
-                    ),
-                });
-            }
-            blobs.push(NodeExecutionCheckpointBlob {
-                content_hash: blob_ref.content_hash.clone(),
-                bytes,
-            });
-        }
-        Ok(Some(NodeExecutionCheckpointBundle {
-            height: descriptor.height,
-            execution_block_hash: descriptor.execution_block_hash.clone(),
-            execution_state_root: descriptor.execution_state_root.clone(),
-            manifest_json,
-            blobs,
-        }))
-    }
-
-    pub(crate) fn pin_execution_checkpoint_descriptor(
-        &self,
-        descriptor: &NodeExecutionCheckpointDescriptor,
-    ) -> Result<(), NodeError> {
-        self.store
-            .pin(descriptor.manifest_ref.as_str())
-            .map_err(distfs_error_to_node_error)?;
-        for blob_ref in &descriptor.blobs {
-            self.store
-                .pin(blob_ref.content_hash.as_str())
-                .map_err(distfs_error_to_node_error)?;
-        }
-        Ok(())
-    }
-
-    pub(crate) fn probe_storage_challenges(
-        &self,
-        world_id: &str,
-        node_id: &str,
-        observed_at_unix_ms: i64,
-    ) -> Result<StorageChallengeProbeReport, NodeError> {
-        let config = StorageChallengeProbeConfig::default();
-        self.store
-            .probe_storage_challenges(world_id, node_id, observed_at_unix_ms, &config)
-            .map_err(distfs_error_to_node_error)
-    }
-
-    pub(crate) fn recent_replicated_content_hashes(
-        &self,
-        world_id: &str,
-        max_samples: usize,
-    ) -> Result<Vec<String>, NodeError> {
-        Ok(self
-            .recent_replicated_content_refs(world_id, max_samples)?
-            .into_iter()
-            .map(|(_, content_hash)| content_hash)
-            .collect())
-    }
-
-    pub(crate) fn recent_replicated_content_refs(
-        &self,
-        world_id: &str,
-        max_samples: usize,
-    ) -> Result<Vec<(u64, String)>, NodeError> {
-        if max_samples == 0 || self.writer_state.last_replicated_height == 0 {
-            return Ok(Vec::new());
-        }
-
-        let mut samples = Vec::with_capacity(max_samples);
-        let mut seen = BTreeSet::new();
-        let mut height = self.writer_state.last_replicated_height;
-        while height > 0 && samples.len() < max_samples {
-            if let Some(message) = self.load_commit_message_by_height(world_id, height)? {
-                let content_hash = message.record.content_hash.trim();
-                if !content_hash.is_empty() && seen.insert(content_hash.to_string()) {
-                    samples.push((height, content_hash.to_string()));
-                }
-            }
-            height -= 1;
-        }
-        Ok(samples)
-    }
-
-    pub(crate) fn replicated_content_refs_from_height(
-        &self,
-        world_id: &str,
-        start_height: u64,
-        max_samples: usize,
-    ) -> Result<Vec<(u64, String)>, NodeError> {
-        if max_samples == 0 || self.writer_state.last_replicated_height == 0 {
-            return Ok(Vec::new());
-        }
-
-        let mut height = start_height.max(1);
-        let latest_height = self.writer_state.last_replicated_height;
-        if height > latest_height {
-            return Ok(Vec::new());
-        }
-
-        let mut samples = Vec::with_capacity(max_samples);
-        let mut seen = BTreeSet::new();
-        while height <= latest_height && samples.len() < max_samples {
-            if let Some(message) = self.load_commit_message_by_height(world_id, height)? {
-                let content_hash = message.record.content_hash.trim();
-                if !content_hash.is_empty() && seen.insert(content_hash.to_string()) {
-                    samples.push((height, content_hash.to_string()));
-                }
-            }
-            height = match height.checked_add(1) {
-                Some(next_height) => next_height,
-                None => break,
-            };
-        }
-        Ok(samples)
     }
 
     fn should_process_remote_message(
